requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
astunparse==1.6.3
attrs==26.1.0
bcrypt==5.0.0
blinker==1.9.0
build==1.4.2
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
chromadb==1.5.5
click==8.3.2
colorama==0.4.6
curl_cffi==0.13.0
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.18.0
durationpy==0.10
filelock==3.25.2
finnhub-python==2.4.27
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
groq==0.37.1
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.9.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.12.0
kubernetes==35.0.0
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-groq==1.1.2
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
libclang==18.1.1
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.8.4
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
multitasking==0.0.12
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.18.1
networkx==3.6.1
numba==0.61.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.24.4
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.7
ormsgpack==1.12.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==26.0
pandas==3.0.2
pandas-ta==0.4.71b0
peewee==3.18.2
pillow==12.1.1
propcache==0.4.1
protobuf==6.33.6
pyarrow==23.0.1
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pypdf==6.9.2
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.4.2
scipy==1.17.1
selenium==4.36.0
sentence-transformers==5.3.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
transformers==5.5.0
trio==0.31.0
trio-websocket==0.12.2
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wsproto==1.2.0
xgboost==3.1.1
xxhash==3.6.0
yarl==1.23.0
yfinance==0.2.66
zipp==3.23.0
zstandard==0.25.0
